Create all busines rules, not validate all

+++ src/main/java/com/curso/tecnologia/dto/PurchaseResponseDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class PurchaseResponseDTO {
     private String customerCpf;
     private ProductDTO product;
     private int quantity;
-    private double totalValue;
+    private BigDecimal totalValue;
 
 }
